Initial transfer of radio buttons
Fixed tests

Moved Files

Fixed coloured outcome

Fixed tests

Created component module

Added validation to non-pass page

Tidy up office page

Disabled activity code on office

Added e2e tests
