+++ src/components/common/common-components.module.ts
@@ -18,11 +18,9 @@ import { TickIndicatorComponent } from './tick-indicator/tick-indicator';
 import { ModalAlertTitleComponent } from './modal-alert-title/modal-alert-title';
 import { ModalReturnButtonComponent } from './modal-return-button/modal-return-button';
 import { ErrorMessageComponent } from './error-message/error-message';
-import { D255Component } from '../../pages/office/components/d255/d255';
-import { DebriefWitnessedComponent } from '../../pages/office/components/debrief-witnessed/debrief-witnessed';
-import { LanguagePreferencesComponent } from '../../pages/office/components/language-preference/language-preferences';
 import { TabComponent } from './tab/tab';
 import { TabsComponent } from './tabs/tabs';
+import { ActivityCodeComponent } from '../../pages/office/components/activity-code/activity-code';
 
 @NgModule({
   declarations: [
@@ -40,11 +38,9 @@ import { TabsComponent } from './tabs/tabs';
     ModalAlertTitleComponent,
     ModalReturnButtonComponent,
     ErrorMessageComponent,
-    LanguagePreferencesComponent,
-    D255Component,
-    DebriefWitnessedComponent,
     TabComponent,
     TabsComponent,
+    ActivityCodeComponent,
   ],
   imports: [
     SignaturePadModule,
@@ -68,11 +64,9 @@ import { TabsComponent } from './tabs/tabs';
     ModalAlertTitleComponent,
     ModalReturnButtonComponent,
     ErrorMessageComponent,
-    LanguagePreferencesComponent,
-    D255Component,
-    DebriefWitnessedComponent,
     TabComponent,
     TabsComponent,
+    ActivityCodeComponent,
   ],
 })
 export class ComponentsModule { }
