--- src/pages/non-pass-finalisation/non-pass-finalisation.html
@@ -1,15 +1,0 @@
-<ion-header>
-  <practice-mode-banner *ngIf="isPracticeMode"></practice-mode-banner>
-  <ion-navbar>
-    <ion-title>Finalise outcome - {{ pageState.candidateName$ | async }}</ion-title>
-  </ion-navbar>
-</ion-header>
-<ion-content no-bounce id="non-pass-finalisation-page">
-  <ion-footer>
-    <ion-row class="mes-full-width-card mes-row-separator">
-      <button type="submit" class="mes-primary-button" ion-button>
-        <h3 id="continue-to-back-to-office" (click)="continue()">Continue</h3>
-      </button>
-    </ion-row>
-  </ion-footer>
-</ion-content>
