--- src/pages/non-pass-finalisation/non-pass-finalisation.scss
@@ -1,2 +1,0 @@
-non-pass-finalisation {
-}
