+++ src/pages/office/__tests__/office.spec.ts
@@ -87,7 +87,6 @@ describe('OfficePage', () => {
         MockComponent(AdditionalInformationComponent),
         MockComponent(IndependentDrivingComponent),
         MockComponent(FaultCommentCardComponent),
-        MockComponent(ActivityCodeComponent),
       ],
       imports: [
         IonicModule,
@@ -351,19 +350,6 @@ describe('OfficePage', () => {
       component.isPracticeMode = true;
       component.popToRoot();
       expect(navController.popTo).toHaveBeenCalled();
-    });
-  });
-
-  describe('isWelshChanged', () => {
-    it('should dispatch a CandidateChoseToProceedInWelshLanguage action when welsh is toggled true', () => {
-      const isWelsh = true;
-      component.isWelshChanged(isWelsh);
-      expect(store$.dispatch).toHaveBeenCalledWith(new CandidateChoseToProceedWithTestInWelsh('Cymraeg'));
-    });
-    it('should dispatch a CandidateChoseToProceedInEnglishLanguage action when welsh is toggled false', () => {
-      const isWelsh = false;
-      component.isWelshChanged(isWelsh);
-      expect(store$.dispatch).toHaveBeenCalledWith(new CandidateChoseToProceedWithTestInEnglish('English'));
     });
   });
