+++ src/pages/office/components/activity-code/activity-code.html
@@ -1,7 +1,7 @@
 <ion-row class="mes-validated-row mes-row-separator" id="activity-code-card" [formGroup]="formGroup">
     <div class="validation-bar" [class.ng-invalid]="invalid"></div>
     <ion-col class="component-label" col-32 align-self-center>
-      <label>Activity Code</label>
+      <label>Activity code</label>
     </ion-col>
     <ion-col align-self-center>
       <ion-row class="spacing-row"></ion-row>
