+++ src/pages/office/office.html
@@ -34,7 +34,7 @@
         <ion-grid class="grid">
           <activity-code id="activity-code-card" [formGroup]="form"
             [activityCodeModel]="pageState.activityCode$ | async" [activityCodeOptions]="activityCodeOptions"
-            (activityCodeChange)="activityCodeChanged($event)"></activity-code>
+            (activityCodeChange)="activityCodeChanged($event)" [disabled]="true"></activity-code>
 
           <div no-padding [hidden]="!(pageState.isTestOutcomeSet$ | async)">
             <route-number [display]="pageState.displayRouteNumber$ | async"
@@ -50,10 +50,6 @@
               [candidateDescription]="pageState.candidateDescription$ | async"
               [outcome]="pageState.testOutcome$ | async"
               (candidateDescriptionChange)="candidateDescriptionChanged($event)"></candidate-description>
-
-            <debrief-witnessed [display]="pageState.displayDebriefWitnessed$ | async" [formGroup]="form"
-              [debriefWitnessed]="pageState.debriefWitnessed$ | async" [outcome]="pageState.testOutcome$ | async"
-              (debriefWitnessedChange)="debriefWitnessedChanged($event)"></debrief-witnessed>
 
             <identification [display]="pageState.displayIdentification$ | async" [formGroup]="form"
               [identification]="pageState.identification$ | async" [outcome]="pageState.testOutcome$ | async"
@@ -87,16 +83,11 @@
               (weatherConditionsChange)="weatherConditionsChanged($event)">
             </weather-conditions>
 
-            <d255 [display]="pageState.displayWeatherConditions$ | async" [d255]="pageState.d255$ | async"
-              [outcome]="pageState.testOutcome$ | async" [formGroup]="form" (d255Change)="d255Changed($event)"></d255>
-
             <additional-information [display]="pageState.displayAdditionalInformation$ | async"
               [additionalInformation]="pageState.additionalInformation$ | async" [formGroup]="form"
               [outcome]="pageState.testOutcome$ | async"
               (additionalInformationChange)="additionalInformationChanged($event)"></additional-information>
 
-            <language-preferences [formGroup]="form" [isWelsh]="pageState.isWelshTest$ | async"
-              (welshChanged)="isWelshChanged($event)"></language-preferences>
           </div>
         </ion-grid>
       </ion-card-content>
