+++ src/pages/test-finalisation/components/d255/d255.html
@@ -1,13 +1,13 @@
 <ion-row *ngIf="display" class="mes-validated-row mes-row-separator" id="d255-card" [formGroup]="formGroup">
   <div class="validation-bar" [class.ng-invalid]="invalid">
   </div>
-  <ion-col class="component-label" col-36 align-self-center>
-    <label>D255?</label>
+  <ion-col class="component-label" col-32 align-self-center>
+    <label>D255</label>
   </ion-col>
   <ion-col align-self-center>
     <ion-row class="spacing-row"></ion-row>
     <ion-row>
-      <ion-col col-32>
+      <ion-col col-36>
         <input type="radio" id="d255-yes" class="gds-radio-button" name="d255" formControlName="d255" value="Yes"
           [class.ng-invalid]="invalid" [checked]="d255 === true" (change)="d255Changed($event.target.value)">
         <label for="d255-yes" class="radio-label">Yes</label>
