+++ src/pages/test-finalisation/components/finalisation-header/finalisation-header.html
@@ -1,0 +1,28 @@
+<ion-grid>
+  <ion-row class="header-card top-row">
+    <ion-col>
+      <h2 id="candidate-name">{{candidateName}}</h2>
+    </ion-col>
+  </ion-row>
+  <ion-row class="header-card bottom-row">
+    <ion-col>
+      <h3 id="candidate-driver-number">{{candidateDriverNumber}}</h3>
+    </ion-col>
+    <ion-col col-36 class="align-col-right">
+      <h3>test outcome: 
+        <div id="test-outcome-text" [ngSwitch]="outcomeText">
+          <div class="pass" *ngSwitchCase="'Passed'">
+            {{outcomeText}}
+          </div>
+          <div class="fail" *ngSwitchCase="'Unsuccessful'">
+            {{outcomeText}}
+          </div>
+          <div class="terminated" *ngSwitchCase="'Terminated'">
+            {{outcomeText}}
+          </div>
+        </div>
+      </h3>
+    </ion-col>
+  </ion-row>
+  <ion-row class="mes-row-separator"></ion-row>
+</ion-grid>
