+++ src/pages/test-finalisation/components/finalisation-header/finalisation-header.ts
@@ -1,0 +1,17 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'finalisation-header',
+  templateUrl: 'finalisation-header.html',
+})
+export class FinalisationHeaderComponent {
+
+  @Input()
+  candidateName: string;
+
+  @Input()
+  candidateDriverNumber: string;
+
+  @Input()
+  outcomeText: string;
+}
