+++ src/pages/test-finalisation/non-pass-finalisation/__tests__/non-pass-finalisation.analytics.effects.spec.ts
@@ -2,17 +2,17 @@ import { TestBed } from '@angular/core/testing';
 import { ReplaySubject } from 'rxjs/ReplaySubject';
 import { StoreModule, Store } from '@ngrx/store';
 import { provideMockActions } from '@ngrx/effects/testing';
-import { AnalyticsProvider } from '../../../providers/analytics/analytics';
-import { AnalyticsProviderMock } from '../../../providers/analytics/__mocks__/analytics.mock';
-import { AnalyticsScreenNames, AnalyticsEventCategories } from '../../../providers/analytics/analytics.model';
-import { AnalyticRecorded } from '../../../providers/analytics/analytics.actions';
-import { StoreModel } from '../../../shared/models/store.model';
-import * as journalActions from '../../journal/journal.actions';
-import * as fakeJournalActions from '../../fake-journal/fake-journal.actions';
-import { testsReducer } from '../../../modules/tests/tests.reducer';
-import { PopulateCandidateDetails } from '../../../modules/tests/candidate/candidate.actions';
+import { AnalyticsProvider } from '../../../../providers/analytics/analytics';
+import { AnalyticsProviderMock } from '../../../../providers/analytics/__mocks__/analytics.mock';
+import { AnalyticsScreenNames, AnalyticsEventCategories } from '../../../../providers/analytics/analytics.model';
+import { AnalyticRecorded } from '../../../../providers/analytics/analytics.actions';
+import { StoreModel } from '../../../../shared/models/store.model';
+import * as journalActions from '../../../journal/journal.actions';
+import * as fakeJournalActions from '../../../fake-journal/fake-journal.actions';
+import { testsReducer } from '../../../../modules/tests/tests.reducer';
+import { PopulateCandidateDetails } from '../../../../modules/tests/candidate/candidate.actions';
 import { Candidate } from '@dvsa/mes-journal-schema';
-import { end2endPracticeSlotId } from '../../../shared/mocks/test-slot-ids.mock';
+import { end2endPracticeSlotId } from '../../../../shared/mocks/test-slot-ids.mock';
 import { NonPassFinalisationAnalyticsEffects } from '../non-pass-finalisation.analytics.effects';
 import * as nonPassFinalisationActions from '../non-pass-finalisation.actions';
