+++ src/pages/test-finalisation/non-pass-finalisation/non-pass-finalisation.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { NonPassFinalisationPage } from './non-pass-finalisation';
-import { AnalyticsProvider } from '../../providers/analytics/analytics';
-import { ComponentsModule } from '../../components/common/common-components.module';
+import { AnalyticsProvider } from '../../../providers/analytics/analytics';
+import { ComponentsModule } from '../../../components/common/common-components.module';
+import { TestFinalisationComponentsModule } from '../components/test-finalisation.module';
 
 @NgModule({
   declarations: [
@@ -11,6 +12,7 @@ import { ComponentsModule } from '../../components/common/common-components.modu
   imports: [
     IonicPageModule.forChild(NonPassFinalisationPage),
     ComponentsModule,
+    TestFinalisationComponentsModule,
   ],
   providers: [
     AnalyticsProvider,
