+++ src/pages/test-finalisation/non-pass-finalisation/non-pass-finalisation.scss
@@ -1,0 +1,49 @@
+non-pass-finalisation {
+  ion-content.content-ios {
+    background-color: map-get($colors, "mes-white");
+  }
+  .mes-primary-button {
+    min-width: 150px;
+  }
+  ion-content {
+    ion-select {
+      width: 452px;
+    }
+    .mes-full-width-card {
+      height: 120px;
+      #continue-to-test-report {
+        width: 100%;
+        height: 56px;
+      }
+    }
+    .manual-label-top-spacing {
+      label {
+        display: block;
+        margin-top: 42px;
+      }
+    }
+  }
+
+  ion-footer {
+    text-align: center;
+    padding: 16px 0;
+    #continue-button {
+      min-height: 56px;
+      width: 746px;
+      margin-top: 11px;
+    }
+  }
+  .header-card {
+    margin-left: 32px;
+  }
+  .top-row {
+    margin-top: 24px;
+  }
+  .bottom-row {
+    margin-bottom: 24px;
+  }
+  #test-outcome {
+    color: map-get($colors, "mes-red") !important;
+    
+  }
+}
