+++ src/pages/test-finalisation/non-pass-finalisation/non-pass-finalisation.ts
@@ -1,0 +1,172 @@
+import { Component } from '@angular/core';
+import { IonicPage, NavController, Platform } from 'ionic-angular';
+import { PracticeableBasePageComponent } from '../../../shared/classes/practiceable-base-page';
+import { AuthenticationProvider } from '../../../providers/authentication/authentication';
+import { Store, select } from '@ngrx/store';
+import { StoreModel } from '../../../shared/models/store.model';
+import { BACK_TO_OFFICE_PAGE } from '../../page-names.constants';
+import { Observable } from 'rxjs/Observable';
+import { getTests } from '../../../modules/tests/tests.reducer';
+import {
+  getCurrentTest,
+  getJournalData,
+  getActivityCode,
+  getTestOutcome,
+  isTestOutcomeSet,
+  getTestOutcomeText,
+} from '../../../modules/tests/tests.selector';
+import { getCandidate } from '../../../modules/tests/candidate/candidate.reducer';
+import {
+  getUntitledCandidateName,
+  getCandidateDriverNumber,
+  formatDriverNumber,
+} from '../../../modules/tests/candidate/candidate.selector';
+import { NonPassFinalisationViewDidEnter } from './non-pass-finalisation.actions';
+import { map, withLatestFrom } from 'rxjs/operators';
+import { getTestSummary } from '../../../modules/tests/test-summary/test-summary.reducer';
+import { isDebriefWitnessed, getD255 } from '../../../modules/tests/test-summary/test-summary.selector';
+import { getTestSlotAttributes } from '../../../modules/tests/test-slot-attributes/test-slot-attributes.reducer';
+import { isWelshTest } from '../../../modules/tests/test-slot-attributes/test-slot-attributes.selector';
+import {
+  ActivityCodeModel,
+  activityCodeModelList,
+} from '../../office/components/activity-code/activity-code.constants';
+import { FormGroup } from '@angular/forms';
+import { SetActivityCode, PersistTests } from '../../../modules/tests/tests.actions';
+import { OutcomeBehaviourMapProvider } from '../../../providers/outcome-behaviour-map/outcome-behaviour-map';
+import { behaviourMap } from '../../office/office-behaviour-map';
+import {
+  DebriefWitnessed,
+  DebriefUnwitnessed,
+  D255Yes,
+  D255No,
+} from '../../../modules/tests/test-summary/test-summary.actions';
+import { WelshTestChanged } from '../../../modules/tests/test-slot-attributes/test-slot-attributes.actions';
+
+interface NonPassFinalisationPageState {
+  candidateName$: Observable<string>;
+  candidateDriverNumber$: Observable<string>;
+  isTestOutcomeSet$: Observable<boolean>;
+  testOutcome$: Observable<string>;
+  testOutcomeText$: Observable<string>;
+  activityCode$: Observable<ActivityCodeModel>;
+  displayDebriefWitnessed$: Observable<boolean>;
+  debriefWitnessed$: Observable<boolean>;
+  displayD255$: Observable<boolean>;
+  d255$: Observable<boolean>;
+  isWelshTest$: Observable<boolean>;
+}
+
+@IonicPage()
+@Component({
+  selector: 'non-pass-finalisation',
+  templateUrl: 'non-pass-finalisation.html',
+})
+export class NonPassFinalisationPage extends PracticeableBasePageComponent {
+
+  pageState: NonPassFinalisationPageState;
+  form: FormGroup;
+  activityCodeOptions: ActivityCodeModel[];
+
+  constructor(
+    store$: Store<StoreModel>,
+    navController: NavController,
+    public platform: Platform,
+    public authenticationProvider: AuthenticationProvider,
+    private outcomeBehaviourProvider: OutcomeBehaviourMapProvider,
+  ) {
+    super(platform, navController, authenticationProvider, store$);
+    this.form = new FormGroup({});
+    this.activityCodeOptions = activityCodeModelList;
+    this.outcomeBehaviourProvider.setBehaviourMap(behaviourMap);
+  }
+
+  ngOnInit() {
+    const currentTest$ = this.store$.pipe(
+      select(getTests),
+      select(getCurrentTest),
+    );
+    this.pageState = {
+      candidateName$: currentTest$.pipe(
+        select(getJournalData),
+        select(getCandidate),
+        select(getUntitledCandidateName),
+      ),
+      candidateDriverNumber$: currentTest$.pipe(
+        select(getJournalData),
+        select(getCandidate),
+        select(getCandidateDriverNumber),
+        map(formatDriverNumber),
+      ),
+      isTestOutcomeSet$: currentTest$.pipe(
+        select(isTestOutcomeSet),
+      ),
+      testOutcome$: currentTest$.pipe(
+        select(getTestOutcome),
+      ),
+      testOutcomeText$: currentTest$.pipe(
+        select(getTestOutcomeText),
+      ),
+      activityCode$: currentTest$.pipe(
+        select(getActivityCode),
+      ),
+      displayDebriefWitnessed$: currentTest$.pipe(
+        select(getTestOutcome),
+        withLatestFrom(currentTest$.pipe(
+          select(getTestSummary),
+          select(isDebriefWitnessed))),
+        map(([outcome, debrief]) =>
+          this.outcomeBehaviourProvider.isVisible(outcome, 'debriefWitnessed', debrief)),
+      ),
+      debriefWitnessed$: currentTest$.pipe(
+        select(getTestSummary),
+        select(isDebriefWitnessed),
+      ),
+      displayD255$: currentTest$.pipe(
+        select(getTestOutcome),
+        withLatestFrom(currentTest$.pipe(
+          select(getTestSummary),
+          select(getD255))),
+        map(([outcome, d255]) =>
+          this.outcomeBehaviourProvider.isVisible(outcome, 'd255', d255)),
+      ),
+      d255$: currentTest$.pipe(
+        select(getTestSummary),
+        select(getD255),
+      ),
+      isWelshTest$: currentTest$.pipe(
+        select(getJournalData),
+        select(getTestSlotAttributes),
+        select(isWelshTest),
+      ),
+    };
+  }
+
+  ionViewDidEnter(): void {
+    this.store$.dispatch(new NonPassFinalisationViewDidEnter());
+  }
+
+  continue() {
+    Object.keys(this.form.controls).forEach(controlName => this.form.controls[controlName].markAsDirty());
+    if (this.form.valid) {
+      this.store$.dispatch(new PersistTests());
+      this.navController.push(BACK_TO_OFFICE_PAGE);
+    }
+  }
+
+  activityCodeChanged(activityCodeModel: ActivityCodeModel) {
+    this.store$.dispatch(new SetActivityCode(activityCodeModel.activityCode));
+  }
+
+  debriefWitnessedChanged(debriefWitnessed: boolean) {
+    this.store$.dispatch(debriefWitnessed ? new DebriefWitnessed() : new DebriefUnwitnessed());
+  }
+
+  d255Changed(d255: boolean): void {
+    this.store$.dispatch(d255 ? new D255Yes() : new D255No());
+  }
+
+  isWelshChanged(isWelsh: boolean) {
+    this.store$.dispatch(new WelshTestChanged(isWelsh));
+  }
+}
