+++ src/pages/test-finalisation/pass-finalisation/pass-finalisation.module.ts
@@ -3,8 +3,9 @@ import { IonicPageModule } from 'ionic-angular';
 import { PassFinalisationPage } from './pass-finalisation';
 import { EffectsModule } from '@ngrx/effects';
 import { PassFinalisationAnalyticsEffects } from './pass-finalisation.analytics.effects';
-import { AnalyticsProvider } from '../../providers/analytics/analytics';
-import { ComponentsModule } from '../../components/common/common-components.module';
+import { AnalyticsProvider } from '../../../providers/analytics/analytics';
+import { ComponentsModule } from '../../../components/common/common-components.module';
+import { TestFinalisationComponentsModule } from '../components/test-finalisation.module';
 
 @NgModule({
   declarations: [
@@ -14,6 +15,7 @@ import { ComponentsModule } from '../../components/common/common-components.modu
     IonicPageModule.forChild(PassFinalisationPage),
     EffectsModule.forFeature([PassFinalisationAnalyticsEffects]),
     ComponentsModule,
+    TestFinalisationComponentsModule,
   ],
   providers: [
     AnalyticsProvider,
